2.5. genLogMessage.gen вместо byte требует int. Новый метод doubleLogWriter.write, который позволяет отправлять сообщение и в консоль, и в логфайл сразу.

+++ src/space/kiritron/pixel/logger/genLogMessage.java
@@ -45,7 +45,7 @@ public class genLogMessage {
      * @param message Сообщение, которое нужно вывести в консоль.
      * @return Возвращает сгенерированное сообщение.
      */
-    public static String gen(byte type, boolean timeWithDate, boolean color, String message) {
+    public static String gen(int type, boolean timeWithDate, boolean color, String message) {
         if (type == 1) {
             if (timeWithDate == false) {
                 return "[" + TADClass.GetTimeWithSeconds + " " + INFO + "]: " + message;
